Backend/requirements.txt
alembic==1.13.1
apns2==0.7.2
APScheduler==3.10.4
blinker==1.7.0
boto3==1.34.33
botocore==1.34.33
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
Flask==3.0.0
Flask-APScheduler==1.13.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
Mako==1.3.0
MarkupSafe==2.1.3
pyapns2==2.0.0
pycparser==2.21
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
tzlocal==5.2
urllib3==1.26.18
Werkzeug==3.0.1
zipp==3.17.0
